Commented and finalized
So this is the final version I was able to achieve. My initial try, and the reference are commented out there.
The issue is that the calculation of q(v|t,u,w,k) refers to a specific cell in the result of logreg.predict_log_proba(vectorized_word). The workaround I found was to use a matrix of num_tags by num_tags-1 (one less column, because "*" isn't seen in training).
Comments where I want to explain. TODOs are possible things to check